requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coinbase==2.1.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
get-all-tickers==1.7
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.6.1
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandas-market-calendars==1.6.1
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.13.0
psycopg2==2.8.6
py==1.10.0
py-postgresql==1.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
PySocks==1.7.1
pytesseract==0.3.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2021.3.17
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
toolz==0.11.1
tqdm==4.60.0
trading-calendars==2.1.1
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.1.0
Werkzeug==1.0.1
